requirements.txt
bottle
python-firebase